refactor(CanvasMode)[OK-89] Add function isActiveItem, remove canvas mode from MenuAction
