+++ frontend/src/utils/fabricCanvasUtils.ts
